# docs/swagger.yaml
definitions:
...
  server.ErrorResponse:
    properties:
      error:
        example: Resource not found
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/ai/attach/{worktree}:
    get:
      description: Establish WebSocket connection for terminal access to AI containers
      parameters:
      - description: Worktree name
        in: path
        name: worktree
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: WebSocket endpoint for AI container terminal
      tags:
      - ai
      - websocket
...
swagger: "2.0"